+++ expectation.messagematch.go
@@ -2,6 +2,8 @@ package testkit
 
 import (
 	"errors"
+	"fmt"
+	"reflect"
 
 	"github.com/dogmatiq/configkit/message"
 	"github.com/dogmatiq/dogma"
@@ -14,22 +16,23 @@ import (
 // ToExecuteCommandMatching returns an expectation that passes if a command is
 // executed that satisfies the given predicate function.
 //
-// Always prefer using ToExecuteCommand() instead, if possible, as it provides
+// Always prefer using [ToExecuteCommand] instead, if possible, as it provides
 // more meaningful information in the result of a failure.
 //
 // pred is the predicate function. It is called for each executed command. It
 // must return nil at least once for the expectation to pass.
 //
-// pred may return the IgnoreMessage error to indicate that the predicate does
-// not apply to a specific message.
-func ToExecuteCommandMatching(
-	pred func(dogma.Command) error,
+// pred may return the [IgnoreMessage] error to indicate that the predicate does
+// not apply to a specific message. Any command that is not of type T is also
+// ignored.
+func ToExecuteCommandMatching[T dogma.Command](
+	pred func(T) error,
 ) Expectation {
 	if pred == nil {
 		panic("ToExecuteCommandMatching(<nil>): function must not be nil")
 	}
 
-	return &messageMatchExpectation{
+	return &messageMatchExpectation[T]{
 		pred:         pred,
 		expectedRole: message.CommandRole,
 		exhaustive:   false,
@@ -43,18 +46,20 @@ func ToExecuteCommandMatching(
 // as recording an event.
 //
 // pred is the predicate function. It is called for each executed command. It
-// must return nil (or IgnoreMessage) every time for the expectation to pass.
+// must return nil (or [IgnoreMessage]) every time for the expectation to pass.
 //
-// pred may return the IgnoreMessage error to indicate that the predicate does
+// pred may return the [IgnoreMessage] error to indicate that the predicate does
 // not apply to a specific message.
-func ToOnlyExecuteCommandsMatching(
-	pred func(dogma.Command) error,
+//
+// If any command is executed that is not of type T, the expectation fails.
+func ToOnlyExecuteCommandsMatching[T dogma.Command](
+	pred func(T) error,
 ) Expectation {
 	if pred == nil {
 		panic("ToOnlyExecuteCommandsMatching(<nil>): function must not be nil")
 	}
 
-	return &messageMatchExpectation{
+	return &messageMatchExpectation[T]{
 		pred:         pred,
 		expectedRole: message.CommandRole,
 		exhaustive:   true,
@@ -64,22 +69,23 @@ func ToOnlyExecuteCommandsMatching(
 // ToRecordEventMatching returns an expectation that passes if an event is
 // recorded that satisfies the given predicate function.
 //
-// Always prefer using ToRecordEvent() instead, if possible, as it provides
-// more meaningful information in the result of a failure.
+// Always prefer using [ToRecordEvent] instead, if possible, as it provides more
+// meaningful information in the result of a failure.
 //
-// pred is the predicate function. It is called for each recorded event. It
-// must return nil at least once for the expectation to pass.
+// pred is the predicate function. It is called for each recorded event. It must
+// return nil at least once for the expectation to pass.
 //
-// pred may return the IgnoreMessage error to indicate that the predicate does
-// not apply to a specific message.
-func ToRecordEventMatching(
-	pred func(dogma.Command) error,
+// pred may return the [IgnoreMessage] error to indicate that the predicate does
+// not apply to a specific message. Any event that is not of type T is also
+// ignored.
+func ToRecordEventMatching[T dogma.Event](
+	pred func(T) error,
 ) Expectation {
 	if pred == nil {
 		panic("ToRecordEventMatching(<nil>): function must not be nil")
 	}
 
-	return &messageMatchExpectation{
+	return &messageMatchExpectation[T]{
 		pred:         pred,
 		expectedRole: message.EventRole,
 		exhaustive:   false,
@@ -93,18 +99,20 @@ func ToRecordEventMatching(
 // as executing a command.
 //
 // pred is the predicate function. It is called for each recorded event. It
-// must return nil (or IgnoreMessage) every time for the expectation to pass.
+// must return nil (or [IgnoreMessage]) every time for the expectation to pass.
 //
-// pred may return the IgnoreMessage error to indicate that the predicate does
+// pred may return the [IgnoreMessage] error to indicate that the predicate does
 // not apply to a specific message.
-func ToOnlyRecordEventsMatching(
-	pred func(dogma.Command) error,
+//
+// If any event is recorded that is not of type T, the expectation fails.
+func ToOnlyRecordEventsMatching[T dogma.Event](
+	pred func(T) error,
 ) Expectation {
 	if pred == nil {
 		panic("ToOnlyRecordEventsMatching(<nil>): function must not be nil")
 	}
 
-	return &messageMatchExpectation{
+	return &messageMatchExpectation[T]{
 		pred:         pred,
 		expectedRole: message.EventRole,
 		exhaustive:   true,
@@ -119,15 +127,16 @@ var IgnoreMessage = errors.New("this message does not need to be inspected by th
 // messageMatchExpectation is an Expectation that checks that at least one
 // message that satisfies a predicate function is produced.
 //
-// It is the implementation used by ToExecuteCommandMatching() and
-// ToRecordEventMatching().
-type messageMatchExpectation struct {
-	pred         func(dogma.Message) error
+// It is the implementation used by [ToExecuteCommandMatching],
+// [ToRecordEventMatching], [ToOnlyExecuteCommandsMatching], and
+// [ToOnlyRecordEventsMatching].
+type messageMatchExpectation[T dogma.Message] struct {
+	pred         func(T) error
 	expectedRole message.Role
 	exhaustive   bool
 }
 
-func (e *messageMatchExpectation) Caption() string {
+func (e *messageMatchExpectation[T]) Caption() string {
 	if e.exhaustive {
 		return inflect.Sprintf(
 			e.expectedRole,
@@ -143,8 +152,8 @@ func (e *messageMatchExpectation) Caption() string {
 	)
 }
 
-func (e *messageMatchExpectation) Predicate(s PredicateScope) (Predicate, error) {
-	return &messageMatchPredicate{
+func (e *messageMatchExpectation[T]) Predicate(s PredicateScope) (Predicate, error) {
+	return &messageMatchPredicate[T]{
 		pred:         e.pred,
 		expectedRole: e.expectedRole,
 		exhaustive:   e.exhaustive,
@@ -155,10 +164,10 @@ func (e *messageMatchExpectation) Predicate(s PredicateScope) (Predicate, error)
 	}, nil
 }
 
-// messageMatchPredicate is the Predicate implementation for
-// messageMatchExpectation.
-type messageMatchPredicate struct {
-	pred         func(dogma.Message) error
+// messageMatchPredicate is the [Predicate] implementation for
+// [messageMatchExpectation].
+type messageMatchPredicate[T dogma.Message] struct {
+	pred         func(T) error
 	expectedRole message.Role
 	exhaustive   bool
 	failures     []*failedMatch
@@ -169,7 +178,7 @@ type messageMatchPredicate struct {
 }
 
 // Notify updates the expectation's state in response to a new fact.
-func (p *messageMatchPredicate) Notify(f fact.Fact) {
+func (p *messageMatchPredicate[T]) Notify(f fact.Fact) {
 	if p.ok {
 		return
 	}
@@ -181,12 +190,19 @@ func (p *messageMatchPredicate) Notify(f fact.Fact) {
 
 // messageProduced updates the predicate's state to reflect the fact that a
 // message of the expected role has been produced.
-func (p *messageMatchPredicate) messageProduced(env *envelope.Envelope) {
+func (p *messageMatchPredicate[T]) messageProduced(env *envelope.Envelope) {
 	if env.Role != p.expectedRole {
 		return
 	}
 
-	err := p.pred(env.Message)
+	var err error
+	if m, ok := env.Message.(T); ok {
+		err = p.pred(m)
+	} else if p.exhaustive {
+		err = fmt.Errorf("predicate function expected %s", reflect.TypeFor[T]())
+	} else {
+		err = IgnoreMessage
+	}
 
 	if err == nil {
 		p.matched++
@@ -198,7 +214,6 @@ func (p *messageMatchPredicate) messageProduced(env *envelope.Envelope) {
 			p.ok = true
 			p.failures = nil
 		}
-
 		return
 	}
 
@@ -224,17 +239,17 @@ func (p *messageMatchPredicate) messageProduced(env *envelope.Envelope) {
 	)
 }
 
-func (p *messageMatchPredicate) Ok() bool {
+func (p *messageMatchPredicate[T]) Ok() bool {
 	return p.ok
 }
 
-func (p *messageMatchPredicate) Done() {
+func (p *messageMatchPredicate[T]) Done() {
 	if p.exhaustive && len(p.failures) == 0 {
 		p.ok = true
 	}
 }
 
-func (p *messageMatchPredicate) Report(treeOk, isInverted bool) *Report {
+func (p *messageMatchPredicate[T]) Report(treeOk, isInverted bool) *Report {
 	rep := &Report{
 		TreeOk: treeOk,
 		Ok:     p.ok,
@@ -292,11 +307,15 @@ func (p *messageMatchPredicate) Report(treeOk, isInverted bool) *Report {
 	reportNoMatch(rep, &p.tracker)
 
 	if p.exhaustive {
+		matched := fmt.Sprintf("only %d of %d", p.matched, p.tracker.produced-p.ignored)
+		if p.matched == 0 {
+			matched = fmt.Sprintf("none of the %d", p.tracker.produced-p.ignored)
+		}
+
 		rep.Explanation = inflect.Sprintf(
 			p.expectedRole,
-			"only %d of %d relevant <messages> matched the predicate",
-			p.matched,
-			p.tracker.produced-p.ignored,
+			"%s relevant <messages> matched the predicate",
+			matched,
 		)
 	}